requirements-hf.txt
gradio>=4.0.0
rembg[cpu]>=2.0.72
Pillow>=10.2.0
